+++ code/system/sys.data.json/src/common/index.ts
@@ -1,5 +1,4 @@
-import type * as t from './types';
+export type * as t from './t';
 
-export type { t };
 export * from '../index.pkg.mjs';
 export * from './libs';
